requirement.txt
Flask==2.3.2
openai==0.27.0
python-dotenv==1.0.0
gunicorn==20.1.0